+++ internal/serve/serve_test.go
@@ -473,6 +473,7 @@ func Test_handleHTTP_authenticatedEndpoints(t *testing.T) {
 		// api-keys
 		{http.MethodPost, "/api-keys"},
 		{http.MethodGet, "/api-keys"},
+		{http.MethodDelete, "/api-keys/12345"},
 	}
 
 	// Expect 401 as a response: